Correciones a vistas en el modelo playlistCancion, asi como en las fu…

+++ src/controller/playlistController.ts
@@ -4,20 +4,24 @@ import prisma from "../prismaclient"
 
 //Crear una playlist asociada a un usuario
 export const crear_playlist = async (req:Request, res:Response):Promise<void> =>{
-    const datos = req.body
-    console.log(datos.name)
-    console.log(datos.userId)
+    const datos = req.body    
     try{
-        const playlist = await prisma.playlist.create({data:{
-            name:datos.name,
-            user:{
-                connect:{ id : Number(datos.userId)}
-            }    
-        }})
-        if (playlist!== null){
-            res.status(201).json(playlist)
-            
+        const user = await prisma.usuario.findUnique({where:{id:Number(datos.userId)}}).then(Boolean)
+        if (user){
+            const playlist = await prisma.playlist.create({data:{
+                name:datos.name,
+                user:{
+                    connect:{ id : Number(datos.userId)}
+                }    
+            }})
+            if (playlist!== null){
+                res.status(201).json(playlist)            
+            }
+
+        }else{
+            res.status(400).json({message:"El usuario no existe"})
         }
+        
         
     }catch(e){
         res.status(400).json({"message":"Ha ocurrido un error en el sevidor intente más tarde"})
@@ -27,68 +31,81 @@ export const crear_playlist = async (req:Request, res:Response):Promise<void> =>
 
 //Obtener playlist y sus canciones
 export const playlist_info = async(req:Request, res:Response):Promise<void> =>{    
-    const playlist_id = Number(req.params.id)
-    console.log(playlist_id)
+    const playlist_id = Number(req.params.id)    
     try{
-        //consultamos la informacion de la playlist
+        //consultamos si existe la Playlist      
         const info_playlist = await prisma.playlist.findUnique({
             where:{
                 id:playlist_id
             }
-        })
-
-        //Encontramos todas las canciones asociadas
-        const datos_playlist_song =  await prisma.playlistSong.findMany({
-            where:{
-            playlistId: playlist_id                
-        }})
-                    
-        //Extraemos el id de las canciones asociadas a la playlist
-        const canciones = datos_playlist_song.map((item)=>item.songId)
-            
-        //Consultamos la data de nuestras canciones
-        const song_info = await prisma.song.findMany({
-            where:{
-                id:{
-                    in:canciones
-                }
-            } 
-        });                       
-        //Armamos nuestro diccionario        
-        if (info_playlist!== null){        
-            let data_json = info_playlist;
-            if (song_info!== null){
-                data_json['songs'] = song_info; 
-            }                                    
-            res.status(201).json(data_json)                    
-        }        
+        })                
+        if(!info_playlist){
+            res.status(400).json({message:"No hay registro de una playlist con el id proporcionado."})
+        }else{
+            //Encontramos todas las canciones asociadas
+            const datos_playlist_song =  await prisma.playlistCancion.findMany({
+                where:{
+                playlistId: playlist_id                
+            }})
+                        
+            //Extraemos el id de las canciones asociadas a la playlist
+            const canciones = datos_playlist_song.map((item:any)=>item.songId)
+                
+            //Consultamos la data de nuestras canciones
+            const song_info = await prisma.cancion.findMany({
+                where:{
+                    id:{
+                        in:canciones
+                    }
+                } 
+            });                       
+            //Armamos nuestro diccionario        
+            if (info_playlist!== null){        
+                let data_json = info_playlist;
+                if (song_info!== null){
+                    data_json['songs'] = song_info; 
+                }                                    
+                res.status(201).json(data_json)                    
+            }  
+        }
         
     }catch(e){        
-        res.status(400).json({"message":"No hay registro de una playlist con el id proporcionado"})
+        res.status(400).json({"message":"Ha ocurrido un error en el servidor."})
     }
 } 
  
 //Añadir canciones a una playlist
 export const añadir_cancion = async(req:Request, res:Response):Promise<void> =>{
-    const datos = req.body
-    console.log(datos)
+    const datos = req.body    
     try{
-        const datos_playlist =  await prisma.playlistSong.create({data:{
-            playlist:{
-                connect:{ id:Number(datos.playlistId)}
-            },
-            song:{
-                connect:{ id:Number(datos.songId)}
+        const playlist_song = await prisma.playlistCancion.findUnique({
+            where:{
+                playlistId_songId:{
+                    playlistId:Number(datos.playlistId),
+                    songId:Number(datos.songId)
+                }
             }
-        }})
-    
-        if (datos_playlist!== null){
-            res.status(201).json(datos_playlist)
-        }
+        }).then(Boolean);        
+        if (playlist_song){
+            res.status(400).json({message:"Usted ya tiene esta cancion en su playlist."})
+        }else{
+            const datos_playlist =  await prisma.playlistCancion.create({data:{
+                playlist:{
+                    connect:{ id:Number(datos.playlistId)}
+                },
+                song:{
+                    connect:{ id:Number(datos.songId)}
+                }
+            }})
+        
+            if (datos_playlist!== null){
+                res.status(201).json(datos_playlist)
+            }
+        }        
 
     }catch(e){
         res.status(400).json({
-            "message":"Intente nuevamente"
+            message:"La cancion que intenta añadir ya no se encuentra disponible."
         });
     }
 }
@@ -98,7 +115,7 @@ export const añadir_cancion = async(req:Request, res:Response):Promise<void> =>
 export const eliminar_cancion = async(req:Request, res:Response):Promise<void> =>{
     const datos = req.body    
     try{
-        const delete_song = await prisma.playlistSong.delete({
+        const delete_song = await prisma.playlistCancion.delete({
             where:{
                 playlistId_songId:{
                     playlistId: Number(datos.playlistId),
@@ -114,31 +131,35 @@ export const eliminar_cancion = async(req:Request, res:Response):Promise<void> =
 
 
 //Eliminar Playlist
-
 export const eliminar_playlist = async(req:Request, res:Response):Promise<void> =>{    
     const id_playlist = Number(req.params.id)       
-    try{
-        //Eliminamos los registros de la tabla M2M
-        const delete_playlist_relation = await prisma.playlistSong.deleteMany({
-            where:{
-                playlistId:id_playlist
-            }
-        })
-        //Luego eliminamos la playlist
-        const delete_playlist = await prisma.playlist.delete({
-            where:{
-                id:id_playlist
-            }
-        })
-        res.status(200).json({
-            "message":"Playlist eliminada correctamente",
-            "playlist":delete_playlist
-        })
-    }catch(e){
-        console.log(e.message)
-        res.status(400).json({
-            "message":"No se ha podido eliminar la playlist intente más tarde."
+    try{        
+        const playlist_exists = await prisma.playlist.findUnique({where:{id:id_playlist}}).then(Boolean)
+        if (playlist_exists){
+            //Luego eliminamos la playlist
+            const delete_playlist = await prisma.playlist.delete({
+                where:{
+                    id:id_playlist
+                }
+            })
+            res.status(200).json({
+                "message":"Playlist eliminada correctamente",
+                "playlist":delete_playlist
+            })
+        }else{
+            res.status(400).json({message:"La playlist que intenta eliminar no existe."})
+        }
+        
+    }catch(e){                
+        res.status(500).json({
+            "message":"No se ha podido eliminar la playlist intente más tarde.",
+            
         })
     }
 
+}
+
+export const datos_m2m = async(req:Request, res:Response):Promise<void> =>{        
+    const data = await prisma.playlistCancion.findMany({})
+    res.json(data)
 }
